requirements.txt
# Core Streamlit dependencies
streamlit>=1.28.0

# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0

# Visualization libraries
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Additional useful libraries
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Optional: For enhanced UI components
streamlit-option-menu>=0.3.6
streamlit-aggrid>=0.3.4
streamlit-echarts>=0.4.0

# Optional: For data processing
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Optional: For machine learning
scikit-learn>=1.3.0

# Optional: For API integrations
fastapi>=0.100.0
uvicorn>=0.23.0

# AI integrations
openai>=1.51.0

# Google Sheets
gspread>=6.1.4
google-auth>=2.36.0

# Development dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# QR (optional)
qrcode[pil]>=7.4
Pillow>=10.4
matplotlib==3.9.2
wordcloud==1.9.3

# Date parsing
python-dateutil>=2.8.2

